+++ datacube_zarr/driver.py
@@ -3,10 +3,10 @@ Zarr Storage driver for ODC
 Supports storage on S3 and Disk
 Should be able to handle hyperspectral data when ready.
 """
-import itertools
+
 from contextlib import contextmanager
 from json.decoder import JSONDecodeError
-from typing import Any, Dict, Generator, List, Optional, Tuple
+from typing import Any, Generator, Optional, Tuple, Union
 
 import numpy as np
 import xarray as xr
@@ -16,14 +16,14 @@ from datacube.utils import geometry
 from datacube.utils.math import num2numpy
 
 from .utils.retry import retry
-from .utils.uris import uri_join, uri_split
+from .utils.uris import uri_split
 from .zarr_io import ZarrIO
 
 PROTOCOL = ['file', 's3']
 FORMAT = 'zarr'
 
 RasterShape = Tuple[int, ...]
-RasterWindow = Tuple[Tuple[int, int]]
+RasterWindow = Tuple[Union[int, Tuple[int, int]], ...]
 
 
 class ZarrDataSource(object):
@@ -169,88 +169,3 @@ class ZarrReaderDriver(object):
 
 def reader_driver_init() -> ZarrReaderDriver:
     return ZarrReaderDriver()
-
-
-class ZarrWriterDriver(object):
-    def __init__(self) -> None:
-        pass
-
-    @property
-    def aliases(self) -> List:
-        return [f'{a} {b}' for a, b in itertools.product([FORMAT], PROTOCOL)]
-
-    @property
-    def format(self) -> str:
-        return FORMAT
-
-    def mk_uri(self, file_path: str, storage_config: dict) -> str:
-        """
-        Constructs a URI from the file_path and storage config.
-
-        :param Path file_path: The file path of the file to be converted into a URI.
-        :param dict storage_config: The dict holding the storage config found in the
-                                    ingest definition.
-        :return: file_path as a URI that the Driver understands.
-        """
-        driver_alias = storage_config['driver']
-        if driver_alias not in self.aliases:
-            raise ValueError(f'Unknown driver alias: {driver_alias}')
-
-        protocol = driver_alias.split()[1]
-        uri = uri_join(protocol, *str(file_path).split("#", 1))
-        return uri
-
-    def write_dataset_to_storage(
-        self,
-        dataset: xr.Dataset,
-        file_uri: str,
-        global_attributes: Optional[dict] = None,
-        variable_params: Optional[dict] = None,
-        storage_config: Optional[dict] = None,
-        **kwargs: str,
-    ) -> Dict:
-        """
-        Persists a xr.DataSet to storage.
-
-        :param xr.Dataset dataset: The xarray.Dataset to be saved
-        :param str file_uri: The file uri
-        :param Dict global_attributes: Global attributes from the product definition.
-        :param Dict variable_params: Variable parameters from the product definition
-        :param Dict storage_config: Storage config from the product definition
-        :return: a dict containing additional metadata to be saved in the DB
-        """
-        # Flattening atributes: Zarr doesn't allow dicts
-        for var_name in dataset.data_vars:
-            data_var = dataset.data_vars[var_name]
-            if 'spectral_definition' in data_var.attrs:
-                spectral_definition = data_var.attrs.pop('spectral_definition', None)
-                data_var.attrs['dc_spectral_definition_response'] = spectral_definition[
-                    'response'
-                ]
-                data_var.attrs['dc_spectral_definition_wavelength'] = spectral_definition[
-                    'wavelength'
-                ]
-
-        # Renaming units: units is a reserved name in Xarray coordinates
-        for var_name in dataset.coords:
-            coord_var = dataset.coords[var_name]
-            if 'units' in coord_var.attrs:
-                units = coord_var.attrs.pop('units', None)
-                coord_var.attrs['dc_units'] = units
-
-        zio = ZarrIO()
-        # Should be a directory but actually get passed a file, which becomes a directory.
-        metadata = zio.save_dataset_to_zarr(
-            uri=file_uri,
-            dataset=dataset,
-            global_attributes=global_attributes,
-            variable_params=variable_params,
-            storage_config=storage_config,
-        )
-
-        # extra metadata to be stored in database
-        return metadata
-
-
-def writer_driver_init() -> ZarrWriterDriver:
-    return ZarrWriterDriver()