+++ docs/about/whats_new.md
@@ -9,13 +9,17 @@
 - Remove deprecated "+init" from "<auth>:<auth_code>" CRS definitions loaded via xarray
 - Update `rasterio` dependency to version >=1.1.5
 - Suppress expected warnings in unit tests
-- CLI script `set_zarr_product_extra_dim.py` for changing a zarr dataset's extra dimension.
-  - Extra dimension is defined in a product definition yaml under "extra_dimension".
+- CLI script `set_zarr_product_extra_dim.py` for changing a zarr dataset's extra dimensions.
+  - Extra dimensions are defined in a product definition yaml under "extra_dimensions".
 - Change `time_idx` to `band_idx` in Zarr driver.
 - Support for list of nodata vals required for 3D zarrs.
 - Add `transform` attribute from rasters to zarr dataset within zarrify.
 - Revert to universal `"band1"` 2D zarr array naming (not `"band01"`, `"band001"`, etc.).
 - Update `xarray` dependency to version >=0.16.2
+- Remove Zarr writer driver.
+  - Creating Zarr datasets via Datacube ingest is no longer supported. The `zarrify` CLI
+    tool can be used to convert from other formats to Zarr prior to indexing.
+- Zarr dimension rename script supports multiple extra dimensions.
 
 
 ## v0.1.1 (02 September 2020)